+++ Tests/Functional/Domain/Repository/EventTypeRepositoryTest.php
@@ -7,9 +7,6 @@ namespace OliverKlee\Seminars\Tests\Functional\Domain\Repository;
 use OliverKlee\Seminars\Domain\Model\EventType;
 use OliverKlee\Seminars\Domain\Model\NullEventType;
 use OliverKlee\Seminars\Domain\Repository\EventTypeRepository;
-use TYPO3\CMS\Core\Information\Typo3Version;
-use TYPO3\CMS\Core\Utility\GeneralUtility;
-use TYPO3\CMS\Extbase\Object\ObjectManager;
 use TYPO3\TestingFramework\Core\Functional\FunctionalTestCase;
 
 /**
@@ -34,12 +31,7 @@ final class EventTypeRepositoryTest extends FunctionalTestCase
     {
         parent::setUp();
 
-        if ((new Typo3Version())->getMajorVersion() >= 11) {
-            $this->subject = GeneralUtility::makeInstance(EventTypeRepository::class);
-        } else {
-            $objectManager = GeneralUtility::makeInstance(ObjectManager::class);
-            $this->subject = $objectManager->get(EventTypeRepository::class);
-        }
+        $this->subject = $this->get(EventTypeRepository::class);
     }
 
     /**