+++ Tests/Unit/Domain/Repository/Registration/RegistrationRepositoryTest.php
@@ -25,8 +25,12 @@ final class RegistrationRepositoryTest extends UnitTestCase
     {
         parent::setUp();
 
-        $objectManagerStub = $this->createMock(ObjectManagerInterface::class);
-        $this->subject = new RegistrationRepository($objectManagerStub);
+        if (\interface_exists(ObjectManagerInterface::class)) {
+            $objectManagerStub = $this->createStub(ObjectManagerInterface::class);
+            $this->subject = new RegistrationRepository($objectManagerStub);
+        } else {
+            $this->subject = new RegistrationRepository();
+        }
     }
 
     /**